Add jointlimits example code accompanying Chapter 10 "Rotational Joint Limits in Quaternion Space" of "Game Engine Gems 3". API changes: In DualVector3.hpp, renamed 'linvel' to 'moment' In Matrix3x3.hpp, renamed 'cross' to 'skew'

+++ guts/BinomialQueue.hpp
@@ -31,7 +31,7 @@ namespace guts
  
     private:
         static const size_t NO_REF = size_t(~0);   
-        static const size_t PARENT_REF = 1 << 31;      
+        static const size_t PARENT_REF = size_t(1 << 31);      
 
         struct Node
         {
